resolves #3 New conditional message feature (#12)
* add roll dialog template

* add message check to the reminders

* move message checks into their own classes

instead of including them in the advantage reminder classes

* add unit test for AttackMessage

* show multiple messages instead of just one

* remove sorting

Sorting the effects by priority was an attempt to repurpose that to just show one message. Probably not a good idea to change what priority was meant for, which was for how to apply multiple effects on the same thing. Also not as necessary since showing them all anyways.

* await the addMessage method

Fixes the unit tests and changed it into the module code as well to make sure we wait for setting up the dialog hook before proceeding. It worked fine without it but better to be safe

* add test for two messages and check that the hook was setup

* add test cases for AbilityCheckMessage

* add unit tests for AbilitySaveMessage

* add unit tests for SkillMessage

* add unit tests for DeathSaveMessage

* add DamageMessage implementation and tests

* add sorting back

* remove roll dialog replacement template

Was going to use this to completely replace the core dialog but went with adding to the existing one instead. Don't need this anymore

* flip priority around

to match the order Foundry applies active effect changes

* small CSS tweaks

Some changes inspired by #9. Increases the padding by 2px to get close to what it had. I still think the smaller sides is better. Increased the margin to 8px. While not as large, it's the size of the padding of the dialog box (window-content class)

* use dialogOptions.id to get the message on the right roll dialog

* add nested options in for Damage message

* update readme and changelog

* fix readme type in MRE section

* add links to issues in changelog

just a convenience

* add MRE compatibility with damage rolls

MRE configures it's damage wrapper with libWrapper as MIXED. That means it always runs after WRAPPED methods like this module uses. If we are using MRE, always process it so we can pass critical and messages down to MRE and then it can decide what to do with them.