#16 #18 update compatibility notes
Did another round of testing with Better Rolls and noted that messages do not work at all. When it queries for advantage it does not use the default template and doesn't pass along any options anyways.
Added more information about Midi as well. Note to self on messages for attack and damage rolls: it doesn't work because Midi doesn't pass along the dialog options