v10 only and switch from lib-wrapper to the new dnd5e hooks (#32)
* remove data uses or switch to system

* use hook for attacks, still async

* move async code out of the preRollAttack hook

and into the addMessageHook

* add a new AdvSourceMixin

to show the source of advantage/disadvantage

* rework preRollAttack a bit

* change saving throws over

* change ability checks over

* change skill checks

* change tool check and death saves

* change damage rolls over

Removed the special part for MRE since it is not being updated for v10.

* change DAE setup to use it's hook

instead of the ready hook

* update tests

* simplify checkArmorStealth assignment

The DAE version that added support was for v8. Now that we're only supporting v10 it's okay to just check that DAE is running. Skip the version check

* setup new hooks directly

instead of inside the init hook

* some cleanup on the main module file

formatting and move fast forwarding debug log message inside the helper function to cut down duplicate code

* add debug message for DAE setup

* remove effect source feature

Add in it's own PR

* make manifest v10 only and remove lib-wrapper

* some moving around and format with Prettier

* move authors to one line too

* update workflow to replace two more things

* forgot to remove return

also put a debug there

* move debug

* update docs